Added legislative jurisdictions and representatives
This is an initial commit to get the data out there, but this is
turning into spaghetti code without their being a reusable data model
object or funciton for instantiating new jurisdictions and elected
officials, so that is definitely a next step. I also need to separate
and reorganize the relationship between US congress legislators and the
state government jurisdiction.