Additions and refactor to include hyperlocal data
Including DC councilmembers here as a proof of concept, but ultimately
hyperlocal data would all be handled externally and pulled in through a
single API call formatted in the same spec. This example is also a
little confusing since it's DC and hyperlocal districts can get
conflated as either city or state with other models - as the sunlight
labs api was considering them state legislative districts i had to
filter that out for now.